feat: 럭키데이 생성 > 활동 선택 버튼 관련 훅으로 로직 분리 (#249)
