+++ src/components/domain/createLuckyDay/selectActivity/SelectActivity.tsx
@@ -1,13 +1,10 @@
-import React, { Fragment, useState } from "react";
+import React, { Fragment } from "react";
 import { useFormContext } from "react-hook-form";
 
 import { activities, CheckIcon } from "assets";
-import type {
-  Activities,
-  ActivitiesServerModel,
-  CreateLuckyDayForm,
-} from "types";
+import type { ActivitiesServerModel, CreateLuckyDayForm } from "types";
 import { ActivityToggle } from "./container";
+import { useSelectActivity } from "./hooks";
 import * as S from "./SelectActivity.styled";
 
 interface SelectActivityProps {
@@ -15,8 +12,6 @@ interface SelectActivityProps {
 }
 
 function SelectActivity({ data }: SelectActivityProps) {
-  const [toggle, setToggle] = useState<string | null>(null);
-
   const { watch, setValue } = useFormContext<CreateLuckyDayForm>();
 
   const actNos = data?.resData.flatMap((activity) =>
@@ -24,41 +19,12 @@ function SelectActivity({ data }: SelectActivityProps) {
   );
   const currentActsUnChecked = watch("acts")?.filter(({ checked }) => !checked);
 
-  const handleToggle = (toggleLabel: string | null): void =>
-    setToggle(toggleLabel);
-
-  const changeIndex = (
-    arr: Activities[] | undefined,
-    idx1: number,
-    idx2: number
-  ): Activities[] => {
-    if (!arr) return [];
-
-    const newArr = [...arr];
-    [newArr[idx1], newArr[idx2]] = [newArr[idx2], newArr[idx1]];
-
-    return newArr;
-  };
-
-  const handleCheckAllBoxes = (): void => {
-    const arr = changeIndex(data?.resData, 2, 3);
-    const actListArr = watch("acts").flatMap((item) => item.actList);
-
-    const acts = arr
-      .map((activity) => ({
-        category: activity.category ?? "",
-        actList:
-          activity.actList.length > 0 && actListArr.length === 0
-            ? activity.actList.map(({ actNo }) => actNo)
-            : [],
-        checked: currentActsUnChecked?.length === 5,
-      }))
-      .filter(({ category }) => category !== "직접 입력");
-
-    if (!acts) return;
-
-    setValue("acts", acts.reverse());
-  };
+  const { toggle, handleToggle, handleCheckAllBoxes } = useSelectActivity({
+    dataRes: data?.resData,
+    currentActsUnChecked,
+    setValue,
+    watch,
+  });
 
   return (
     <>
