+++ src/components/domain/createLuckyDay/selectActivity/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useSelectActivity } from "./useSelectActivity";
